requirements.txt
python-telegram-bot
openpyxl
python-dotenv
pandas
country-list
